Moved filterBenefits to its own selector
