+++ components/BB.js
@@ -15,6 +15,7 @@ import BenefitList from "../components/benefit_list";
 import NeedsSelector from "./needs_selector";
 import ProfileSelector from "./profile_selector";
 import { connect } from "react-redux";
+import { getFilteredBenefits } from "../selectors";
 
 const styles = theme => ({
   benefitsCount: {
@@ -105,111 +106,6 @@ export class BB extends Component {
     });
   };
 
-  eligibilityMatch = (path, selected) => {
-    let matches = true;
-    ["serviceType", "patronType", "statusAndVitals"].forEach(criteria => {
-      if (
-        selected[criteria] != "" &&
-        path[criteria] !== "na" &&
-        selected[criteria] != path[criteria]
-      ) {
-        matches = false;
-      }
-    });
-    return matches;
-  };
-
-  filterBenefits = (
-    benefits,
-    eligibilityPaths,
-    selectedEligibility,
-    needs,
-    selectedNeeds
-  ) => {
-    if (benefits.length === 0) {
-      return benefits;
-    }
-
-    // make it easy to invert the id
-    let benefitForId = {};
-    benefits.forEach(b => {
-      benefitForId[b.id] = b;
-    });
-
-    // find benefits that match
-    let eligibleBenefitIds = [];
-    eligibilityPaths.forEach(ep => {
-      if (this.eligibilityMatch(ep, selectedEligibility)) {
-        eligibleBenefitIds = eligibleBenefitIds.concat(ep.benefits);
-      }
-    });
-    const eligibleBenefits = eligibleBenefitIds.map(id => benefitForId[id]);
-
-    let benefitIdsForSelectedNeeds = [];
-    if (Object.keys(selectedNeeds).length > 0) {
-      Object.keys(selectedNeeds).forEach(id => {
-        const need = needs.filter(n => n.id === id)[0];
-        benefitIdsForSelectedNeeds = benefitIdsForSelectedNeeds.concat(
-          need.benefits
-        );
-      });
-    } else {
-      benefitIdsForSelectedNeeds = benefits.map(b => b.id);
-    }
-    let matchingBenefitIds = eligibleBenefitIds.filter(
-      id => benefitIdsForSelectedNeeds.indexOf(id) > -1
-    );
-    const matchingBenefits = matchingBenefitIds.map(id => benefitForId[id]);
-
-    /* show:
-         - matching benefits
-         - eligible benefits with a matching child
-         ( maybe: - non-eligible benefits with a matching child that isn't already covered)
-     */
-
-    const eligibleBenefitsWithMatchingChild = eligibleBenefits.filter(
-      b =>
-        b.childBenefits
-          ? b.childBenefits.filter(id => matchingBenefitIds.indexOf(id) > -1)
-              .length > 0
-          : false
-    );
-
-    let benefitsToShow = matchingBenefits.concat(
-      eligibleBenefitsWithMatchingChild
-    );
-    benefitsToShow = benefitsToShow.filter(
-      (b, n) => benefitsToShow.indexOf(b) === n
-    ); // dedup
-
-    // if a benefit is already shown as a child, only show it (as a parent card) if it's available independently
-    let childrenIDsShown = [];
-    benefitsToShow.forEach(b => {
-      childrenIDsShown = childrenIDsShown.concat(b.childBenefits);
-    });
-    benefitsToShow = benefitsToShow.filter(
-      b =>
-        b.availableIndependently === "Independent" ||
-        childrenIDsShown.indexOf(b.id) < 0
-    );
-
-    // If there is a searchString the run another filter
-    if (this.state.searchString.trim() !== "") {
-      let results = [];
-      if (this.props.t("current-language-code") == "en") {
-        results = this.state.enIdx.search(this.state.searchString + "*");
-      } else {
-        results = this.state.frIdx.search(this.state.searchString + "*");
-      }
-      let resultIds = results.map(r => r.ref);
-      benefitsToShow = benefitsToShow.filter(benefit =>
-        resultIds.includes(benefit.id)
-      );
-    }
-
-    return benefitsToShow;
-  };
-
   handleSortByChange = event => {
     this.setState({ sortByValue: event.target.value });
   };
@@ -286,13 +182,7 @@ export class BB extends Component {
   render() {
     const { t, classes } = this.props; // eslint-disable-line no-unused-vars
 
-    const filteredBenefits = this.filterBenefits(
-      this.props.benefits,
-      this.props.eligibilityPaths,
-      this.props.selectedEligibility,
-      this.props.needs,
-      this.props.selectedNeeds
-    );
+    const filteredBenefits = this.props.filteredBenefits;
 
     const printUrl = this.getPrintUrl(
       filteredBenefits,
@@ -433,6 +323,7 @@ const mapStateToProps = reduxState => {
     benefits: reduxState.benefits,
     eligibilityPaths: reduxState.eligibilityPaths,
     examples: reduxState.examples,
+    filteredBenefits: getFilteredBenefits(reduxState),
     needs: reduxState.needs,
     selectedEligibility: {
       patronType: reduxState.patronType,
@@ -448,6 +339,7 @@ BB.propTypes = {
   classes: PropTypes.object.isRequired,
   eligibilityPaths: PropTypes.array.isRequired,
   examples: PropTypes.array.isRequired,
+  filteredBenefits: PropTypes.array,
   id: PropTypes.string.isRequired,
   needs: PropTypes.array.isRequired,
   selectedEligibility: PropTypes.object.isRequired,
