+++ package.json
@@ -53,6 +53,7 @@
     "redux-devtools-extension": "^2.13.2",
     "redux-mock-store": "^1.5.3",
     "redux-thunk": "^2.2.0",
+    "reselect": "^3.0.1",
     "universal-cookie": "^2.2.0",
     "wait-on": "^2.1.0"
   },
